Specify PHP versions explicitly
That way we *know* which version is actually installed, instead of the versions depending on _when_ the layers were built.

That will also help with the general flow: contributors can send PRs to increase PHP versions when they come out. Instead of randomly having to rebuild layers in the hopes that it will bundle a new PHP version.